fix: Jwt 클래스와 Claims 클래스 분리
